Remove all share pages/layouts
