--- src/app/shot-on/[make]/[model]/[photoId]/share/page.tsx
@@ -1,19 +1,0 @@
-import { getPhotoCached } from '@/photo/cache';
-import { PhotoCameraProps, cameraFromPhoto } from '@/camera';
-import PhotoShareModal from '@/photo/PhotoShareModal';
-import { PATH_ROOT } from '@/site/paths';
-import { redirect } from 'next/navigation';
-
-export default async function Share({
-  params,
-}: PhotoCameraProps) {
-  const { photoId, make, model } = await params;
-
-  const photo = await getPhotoCached(photoId);
-
-  if (!photo) { return redirect(PATH_ROOT); }
-
-  const camera = cameraFromPhoto(photo, { make, model });
-
-  return <PhotoShareModal {...{ photo, camera }} />;
-}
